# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger definition for MedRoom backend.",
        "title": "MedRoom API",
        "termsOfService": "https://swagger.io/terms/",
        "contact": {
            "name": "Martín Saavedra",
            "url": "https://www.swagger.io/support",
            "email": "martin.saavedra@mail.udp.cl"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/estudiantes": {
            "get": {
                "description": "Lista todos los estudiantes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Lista de estudiantes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SwaggerMessages.ListEstudiantesSwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Genera un nuevo estudiante con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Agrega un nuevo estudiante",
                "parameters": [
                    {
                        "description": "Estudiante a agregar",
                        "name": "input_estudiante",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestMessages.AddNewEstudiantePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.AddNewEstudianteSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            }
        },
        "/estudiantes/me": {
            "get": {
                "description": "Obtiene el perfil del estudiante según su token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Obtiene el perfil del estudiante",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.GetMyEstudianteSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Modifica el perfil del propio estudiante con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Modifica mi perfil",
                "parameters": [
                    {
                        "description": "Nuevos datos del estudiante a modificar",
                        "name": "input_actualiza_estudiante",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestMessages.PutMyEstudiantePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.PutMyEstudianteSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/estudiantes/{uuid_estudiante}": {
            "get": {
                "description": "Obtiene un estudiante según su UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Obtiene un estudiante",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID del estudiante a buscar",
                        "name": "uuid_estudiante",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.GetOneEstudianteSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Modifica un estudiante con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Modifica un estudiante",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID del estudiante a modificar",
                        "name": "uuid_estudiante",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Estudiante a modificar",
                        "name": "input_actualiza_estudiante",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestMessages.PutOneEstudiantePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.PutOneEstudianteSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un estudiante con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Elimina un estudiante",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID del estudiante a eliminar",
                        "name": "uuid_estudiante",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.DeleteEstudianteSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ApiHelpers.ResponseError": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "RequestMessages.AddNewEstudiantePayload": {
            "type": "object",
            "properties": {
                "apellidos_estudiante": {
                    "type": "string"
                },
                "correo_electronico_estudiante": {
                    "type": "string"
                },
                "hash_contrasena_estudiante": {
                    "type": "string"
                },
                "id_grupo": {
                    "type": "integer"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_estudiante": {
                    "type": "string"
                },
                "rut_estudiante": {
                    "type": "string"
                },
                "telefono_celular_estudiante": {
                    "type": "string"
                },
                "telefono_fijo_estudiante": {
                    "type": "string"
                }
            }
        },
...
        "RequestMessages.PutMyEstudiantePayload": {
            "type": "object",
            "properties": {
                "apellidos_estudiante": {
                    "type": "string"
                },
                "correo_electronico_estudiante": {
                    "type": "string"
                },
                "hash_contrasena_estudiante": {
                    "description": "solo este",
                    "type": "string"
                },
                "id_grupo": {
                    "type": "integer"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_estudiante": {
                    "type": "string"
                },
                "rut_estudiante": {
                    "type": "string"
                },
                "telefono_celular_estudiante": {
                    "description": "solo este",
                    "type": "string"
                },
                "telefono_fijo_estudiante": {
                    "description": "solo este",
                    "type": "string"
                }
            }
        },
...
        "RequestMessages.PutOneEstudiantePayload": {
            "type": "object",
            "properties": {
                "apellidos_estudiante": {
                    "type": "string"
                },
                "correo_electronico_estudiante": {
                    "type": "string"
                },
                "hash_contrasena_estudiante": {
                    "type": "string"
                },
                "id_grupo": {
                    "type": "integer"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_estudiante": {
                    "type": "string"
                },
                "rut_estudiante": {
                    "type": "string"
                },
                "telefono_celular_estudiante": {
                    "type": "string"
                },
                "telefono_fijo_estudiante": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.AddNewEstudianteResponse": {
            "type": "object",
            "properties": {
                "apellidos_estudiante": {
                    "type": "string"
                },
                "correo_electronico_estudiante": {
                    "type": "string"
                },
                "id_grupo": {
                    "type": "integer"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_estudiante": {
                    "type": "string"
                },
                "rut_estudiante": {
                    "type": "string"
                },
                "telefono_celular_estudiante": {
                    "type": "string"
                },
                "telefono_fijo_estudiante": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.DeleteEstudianteResponse": {
            "type": "object",
            "properties": {
                "apellidos_estudiante": {
                    "type": "string"
                },
                "correo_electronico_estudiante": {
                    "type": "string"
                },
                "id_grupo": {
                    "type": "integer"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_estudiante": {
                    "type": "string"
                },
                "rut_estudiante": {
                    "type": "string"
                },
                "telefono_celular_estudiante": {
                    "type": "string"
                },
                "telefono_fijo_estudiante": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.GetMyEstudianteResponse": {
            "type": "object",
            "properties": {
                "apellidos_estudiante": {
                    "type": "string"
                },
                "correo_electronico_estudiante": {
                    "type": "string"
                },
                "evaluaciones_estudiante": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessages.ListEvaluacionesEstudianteResponse"
                    }
                },
                "id_grupo": {
                    "type": "integer"
                },
                "nombres_estudiante": {
                    "type": "string"
                },
                "rol_estudiante": {
                    "$ref": "#/definitions/ResponseMessages.GetOneRolResponse"
                },
                "rut_estudiante": {
                    "type": "string"
                },
                "telefono_celular_estudiante": {
                    "type": "string"
                },
                "telefono_fijo_estudiante": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.GetOneEstudianteResponse": {
            "type": "object",
            "properties": {
                "apellidos_estudiante": {
                    "type": "string"
                },
                "correo_electronico_estudiante": {
                    "type": "string"
                },
                "evaluaciones_estudiante": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessages.ListEvaluacionesEstudianteResponse"
                    }
                },
                "id_grupo": {
                    "type": "integer"
                },
                "nombres_estudiante": {
                    "type": "string"
                },
                "rol_estudiante": {
                    "$ref": "#/definitions/ResponseMessages.GetOneRolResponse"
                },
                "rut_estudiante": {
                    "type": "string"
                },
                "telefono_celular_estudiante": {
                    "type": "string"
                },
                "telefono_fijo_estudiante": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.GetOneRolResponse": {
            "type": "object",
            "properties": {
                "nombre_rol": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.ListEstudiantesResponse": {
            "type": "object",
            "properties": {
                "apellidos_estudiante": {
                    "type": "string"
                },
                "correo_electronico_estudiante": {
                    "type": "string"
                },
                "evaluaciones_estudiante": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessages.ListEvaluacionesEstudianteResponse"
                    }
                },
                "id_grupo": {
                    "type": "integer"
                },
                "nombres_estudiante": {
                    "type": "string"
                },
                "rol_estudiante": {
                    "$ref": "#/definitions/ResponseMessages.GetOneRolResponse"
                },
                "rut_estudiante": {
                    "type": "string"
                },
                "telefono_celular_estudiante": {
                    "type": "string"
                },
                "telefono_fijo_estudiante": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.PutMyEstudianteResponse": {
            "type": "object",
            "properties": {
                "apellidos_estudiante": {
                    "type": "string"
                },
                "correo_electronico_estudiante": {
                    "type": "string"
                },
                "hash_contrasena_estudiante": {
                    "type": "string"
                },
                "id_grupo": {
                    "type": "integer"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_estudiante": {
                    "type": "string"
                },
                "rut_estudiante": {
                    "type": "string"
                },
                "telefono_celular_estudiante": {
                    "type": "string"
                },
                "telefono_fijo_estudiante": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.PutOneEstudianteResponse": {
            "type": "object",
            "properties": {
                "apellidos_estudiante": {
                    "type": "string"
                },
                "correo_electronico_estudiante": {
                    "type": "string"
                },
                "hash_contrasena_estudiante": {
                    "type": "string"
                },
                "id_grupo": {
                    "type": "integer"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_estudiante": {
                    "type": "string"
                },
                "rut_estudiante": {
                    "type": "string"
                },
                "telefono_celular_estudiante": {
                    "type": "string"
                },
                "telefono_fijo_estudiante": {
                    "type": "string"
                }
            }
        },
...
        "SwaggerMessages.AddNewEstudianteSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.AddNewEstudianteResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.DeleteEstudianteSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.DeleteEstudianteResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.GetMyEstudianteSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.GetMyEstudianteResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.GetOneEstudianteSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.GetOneEstudianteResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.ListEstudiantesSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.ListEstudiantesResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.PutMyEstudianteSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.PutMyEstudianteResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.PutOneEstudianteSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.PutOneEstudianteResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    }
}